refacto 2 time the project for have the difference between inheritance and composition
