--- csv_reader/src/CsvFileReader.backup.ts
@@ -1,39 +1,0 @@
-import fs from 'fs'; // Node std lib File System.
-
-import { dateStringToDate } from './utils';
-
-import { MatchResult } from './MatchResult';
-
-// Custom Tuples for matching the row of a game with the propers types.
-type MatchData = [Date, string, string, number, number, MatchResult, string];
-
-export class CsvFileReader {
-  data: MatchData[] = [];
-
-  constructor(public filename: string) { }
-
-  read(): void {
-    this.data = fs
-      .readFileSync(this.filename, {
-        encoding: 'utf8',
-        flag: 'r+'
-      })
-      .split('\n')
-      .map(
-        (row: string): string[] => {
-          return row.split(',');
-        }
-      )
-      .map((row: string[]): MatchData => {
-        return [
-          dateStringToDate(row[0]),
-          row[1],
-          row[2],
-          parseInt(row[3]),
-          parseInt(row[4]),
-          row[5] as MatchResult, // 'H', 'A', 'D' Type Insertion
-          row[6]
-        ];
-      });
-  };
-};
