+++ csv_reader/src/composition/MatchReader.ts
@@ -1,0 +1,33 @@
+import { dateStringToDate } from '../utils';
+import { MatchResult } from '../MatchResult';
+
+// Custom Tuples for matching the row of a game with the propers types.
+type MatchData = [Date, string, string, number, number, MatchResult, string];
+
+interface IDataReader {
+  read(): void,
+  data: string[][]
+}
+
+export class MatchReader {
+  matches: MatchData[] = [];
+
+  constructor(public reader: IDataReader) { }
+
+  load(): void {
+    this.reader.read()
+    this.matches = this.reader.data.map(
+      (row: string[]): MatchData => {
+        return [
+          dateStringToDate(row[0]),
+          row[1],
+          row[2],
+          parseInt(row[3]),
+          parseInt(row[4]),
+          row[5] as MatchResult, // 'H', 'A', 'D' Type Insertion
+          row[6]
+        ];
+      }
+    );
+  }
+}
