+++ csv_reader/src/index.ts
@@ -1,13 +1,16 @@
-import { MatchReader } from './MatchReader';
+import { MatchReader } from './composition/MatchReader';
+import { CsvFileReader } from './composition/CsvFileReader';
 import { MatchResult } from './MatchResult';
 
-const reader = new MatchReader('football.csv');
-reader.read();
+const csvFileReader = new CsvFileReader('football.csv');
+const matchReader = new MatchReader(csvFileReader);
+
+matchReader.load();
 
 let manUnitedWins = 0;
 let manUnitedDraw = 0;
 
-for (let match of reader.data) {
+for (let match of matchReader.matches) {
   if (match[1] === 'Man United' && match[5] === MatchResult.HomeWin) {
     manUnitedWins++;
   } else if (match[2] === 'Man United' && match[5] === MatchResult.AwayWin) {
