+++ csv_reader/src/inheritance/MatchReader.ts
@@ -1,6 +1,6 @@
 import { CsvFileReader } from './CsvFileReader'
-import { dateStringToDate } from './utils';
-import { MatchResult } from './MatchResult';
+import { dateStringToDate } from '../utils';
+import { MatchResult } from '../MatchResult';
 
 // Custom Tuples for matching the row of a game with the propers types.
 type MatchData = [Date, string, string, number, number, MatchResult, string];
